# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Chatbot Law API",
    "version": "1.0.0",
    "description": "API documentation for the AI Chatbot for Education Laws."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/legal-docs": {
      "get": {
        "summary": "[USER/ADMIN] Get all legal documents",
        "description": "User và admin đều có thể xem danh sách văn bản pháp luật.",
        "tags": ["LegalDocument"],
        "responses": { "200": { "description": "List of legal documents" } }
      },
      "post": {
        "summary": "[ADMIN] Create a new legal document",
        "description": "Chỉ admin mới được tạo mới văn bản pháp luật.",
        "tags": ["LegalDocument"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "document_name": { "type": "string" },
                  "document_type": { "type": "string" },
                  "document_date_issue": { "type": "string", "format": "date" },
                  "document_signee": { "type": "string" }
                },
                "required": [
                  "document_name",
                  "document_type",
                  "document_date_issue"
                ]
              }
            }
          }
        },
        "responses": { "201": { "description": "Legal document created" } }
      }
    },
    "/legal-docs/upload": {
      "post": {
        "summary": "[ADMIN] Upload PDF, extract and save clauses",
        "description": "Chỉ admin mới được sử dụng. Upload file PDF luật, hệ thống tự động nhận diện dạng text hoặc scan (ảnh). Nếu là scan, sẽ tự động nhận diện tiếng Việt bằng OCR và trích xuất điều khoản.",
        "tags": ["LegalDocument"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload và trích xuất thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "count": { "type": "integer" },
                    "document": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "type": { "type": "string" },
                        "date": { "type": "string", "format": "date-time" }
                      }
                    },
                    "clauses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "number": { "type": "string" },
                          "title": { "type": "string" },
                          "content": { "type": "string" }
                        }
                      }
                    },
                    "markdown": { "type": "string" },
                    "json": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No file uploaded hoặc không trích xuất được điều khoản"
          },
          "500": {
            "description": "Lỗi xử lý file hoặc trích xuất (bao gồm cả lỗi OCR)",
            "content": {
              "application/json": {
                "examples": {
                  "ocrError": {
                    "summary": "Lỗi OCR hoặc chuyển PDF sang ảnh",
                    "value": {
                      "success": false,
                      "error": "Không lấy được buffer ảnh từ pdf2pic. Kiểm tra lại phiên bản pdf2pic hoặc option truyền vào."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/legal-docs/export-files": {
      "post": {
        "summary": "[ADMIN] Upload PDF and export files (no database save)",
        "description": "Upload file PDF, trích xuất điều khoản và xuất ra file Markdown/JSON. Không lưu vào database.",
        "tags": ["LegalDocument"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Xuất file thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "count": { "type": "integer" },
                    "files": {
                      "type": "object",
                      "properties": {
                        "markdown": {
                          "type": "object",
                          "properties": {
                            "content": { "type": "string" },
                            "filename": { "type": "string" }
                          }
                        },
                        "json": {
                          "type": "object",
                          "properties": {
                            "content": { "type": "string" },
                            "filename": { "type": "string" }
                          }
                        },
                        "raw": {
                          "type": "object",
                          "properties": {
                            "content": { "type": "string" },
                            "filename": { "type": "string" }
                          }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No file uploaded hoặc không trích xuất được điều khoản"
          },
          "500": { "description": "Lỗi xử lý file hoặc trích xuất" }
        }
      }
    },
    "/legal-docs/{id}": {
      "get": {
        "summary": "[USER/ADMIN] Get legal document by ID",
        "description": "User và admin đều có thể xem chi tiết văn bản.",
        "tags": ["LegalDocument"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Legal document found" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "[ADMIN] Update legal document by ID",
        "description": "Chỉ admin mới được cập nhật văn bản.",
        "tags": ["LegalDocument"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "document_name": { "type": "string" },
                  "document_type": { "type": "string" },
                  "document_date_issue": { "type": "string", "format": "date" },
                  "document_signee": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Legal document updated" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "[ADMIN] Delete legal document by ID",
        "description": "Chỉ admin mới được xóa văn bản.",
        "tags": ["LegalDocument"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Legal document deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/legal-docs/clauses/{documentId}": {
      "get": {
        "summary": "[USER/ADMIN] Get all clauses of a legal document",
        "description": "Lấy tất cả điều khoản của một văn bản pháp luật, sắp xếp theo số điều.",
        "tags": ["LegalDocument"],
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách điều khoản của văn bản",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "count": { "type": "integer" },
                    "clauses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "number": { "type": "string" },
                          "content": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "description": "Lỗi server" }
        }
      }
    },
    "/legal-docs/export/{documentId}": {
      "get": {
        "summary": "[USER/ADMIN] Export document from database to Markdown/JSON",
        "description": "Xuất văn bản từ database ra file Markdown và JSON.",
        "tags": ["LegalDocument"],
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Xuất document thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "count": { "type": "integer" },
                    "document": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "type": { "type": "string" },
                        "date": { "type": "string", "format": "date-time" }
                      }
                    },
                    "files": {
                      "type": "object",
                      "properties": {
                        "markdown": {
                          "type": "object",
                          "properties": {
                            "content": { "type": "string" },
                            "filename": { "type": "string" }
                          }
                        },
                        "json": {
                          "type": "object",
                          "properties": {
                            "content": { "type": "string" },
                            "filename": { "type": "string" }
                          }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "Document not found" },
          "400": { "description": "No clauses found for this document" },
          "500": { "description": "Lỗi server" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token vào đây. Ví dụ: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      }
    },
...
  }
}